requirements.txt
websocket-client
orjson